Massive restructuring. Established class files, added up to date work, updated my writeup and added scripted the new collision handler. Also added a new case in the app.py to showcase error.
